requirements_bak.txt
alabaster==0.7.12
asgiref==3.5.2
astroid==2.12.13
Babel==2.11.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
dill==0.3.6
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
docopt==0.6.2
docutils==0.19
idna==3.4
imagesize==1.4.1
isort==5.11.4
Jinja2==3.1.2
lazy-object-proxy==1.8.0
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
mysqlclient==2.1.1
packaging==22.0
pipreqs==0.4.11
platformdirs==2.6.2
Pygments==2.13.0
pylint==2.15.9
pytz==2022.7
PyYAML==6.0
requests==2.28.1
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinx-autoapi==2.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
tzdata==2022.7
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.13
wrapt==1.14.1
yarg==0.1.9
